[ 🔧 | #47 | Refactor ] 리팩토링
